+++ api/manager.js
@@ -9,6 +9,7 @@ const TagDao = require('../dao/tagDao.js');
 const UserDao = require('../dao/userDao.js');
 const ScopeDao = require('../dao/scopeDao.js');
 const RoleDao = require('../dao/roleDao.js');
+const SysParamDao = require('../dao/SysParamDao.js');
 const UserAllowDeviceDao = require('../dao/userAllowDeviceDao.js');
 
 let sequelize = null;
@@ -19,7 +20,7 @@ function _getConn () {
 
 function _init (postgresConf) {
   sequelize = new Sequelize(postgresConf.database, postgresConf.username, postgresConf.password, {
-    host: postgresConf.hostname,
+    host: postgresConf.host,
     dialect: 'postgres',
     pool: {
       max: 10,
@@ -27,7 +28,7 @@ function _init (postgresConf) {
       idle: 10000
     },
     quoteIdentifiers: false,
-    logging: false
+    logging: true
   });
 
   ProjectDao.init(sequelize);
@@ -37,6 +38,7 @@ function _init (postgresConf) {
   UserDao.init(sequelize);
   ScopeDao.init(sequelize);
   RoleDao.init(sequelize);
+  SysParamDao.init(sequelize);
   UserAllowDeviceDao.init(sequelize);
 }
 
@@ -50,5 +52,6 @@ module.exports = {
   UserDao: UserDao,
   ScopeDao: ScopeDao,
   RoleDao: RoleDao,
+  SysParamDao: SysParamDao,
   UserAllowDeviceDao: UserAllowDeviceDao
 };